[seller] refactor(204):drag 함수를 hook으로 이동
